requirements.txt
Flask==3.1.0
PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography>=44.0.0
qrcode==7.4.2
Pillow>=10.0.0
gunicorn>=22.0.0

